[FIX]: searchpage viewmodel 분리
