+++ lib/provider.dart
@@ -13,7 +13,7 @@ import 'package:provider/single_child_widget.dart';
 import 'package:alpha_fe/pages/home_page/home_page_view_model/home_page_view_model.dart';
 import 'package:alpha_fe/pages/add_page/add_page_0/view_model/add_page_0_view_model.dart';
 import 'package:alpha_fe/components/appbars/search_appbar/search_appbar_view_model.dart';
-
+import 'package:alpha_fe/pages/add_page/add_page_2/searchplace_page/searchplace_view_model.dart';
 
 
 List<SingleChildWidget> appProviders = [
@@ -28,4 +28,5 @@ List<SingleChildWidget> appProviders = [
   ChangeNotifierProvider(create: (_) => ShowFinalTourViewModel()),
   ChangeNotifierProvider(create: (_) => AddUserViewModel()),
   ChangeNotifierProvider(create: (_) => SearchAppBarViewModel()),
+  ChangeNotifierProvider(create: (_) => SearchPlaceViewModel()),
 ];
